feat: Traverse curried lambda arguments in AST to extract arg names
Adds logic to traverse the AST parsed by rnix for all set entries that
are curried lambdas and collect the argument names.

Currently there is no syntax for attaching doc strings to individual
arguments, but the DocBook output requires a doc string to be present.
This means that all function arguments are currently documented with
the string "Function argument".

v0v

Fixes #4